+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Enhancements
 
+* New `reset` plugin added
 * New `merge` plugin added
 * Improved support for plugins in modules.
 * CONFIGURABLE: Added before_update and after_update callbacks